requirements.txt
streamlit
python-docx
pdfplumber
watchdog